Added a missing 'ref' clause in a forall.
While there, brought crash1callDestructorsAddon.chpl
in sync with crash1callDestructorsMain.chpl.
The former started out as a partial copy of the latter.
While it is not important that the two be identical,
keeping them that way will avoid the question
"why are they different".

Also updated the corresponding .future.