Adjusted .bad for autoCopyWithBegin.chpl.
Probably because of the forall intent commit #614,
this test is now failing earlier.
I updated .bad, and correspondingly the .prediff.
I also have the .prediff remove the chpl version, to avoid noise.

I verified that the "before" and the "after" failures are due to
the same "number of actuals does not match number of formals"
internal error for a call to the same function with the same argument,
at the same location: autoCopyWithBegin.chpl:15.
So it is sufficient to just adjusting .bad.

Recording this just in case - before today the failure was here:

#3  0x00000000004b69f0 in checkFormalActualTypesMatch () at checks.cpp:662
#4  0x00000000004b5d50 in check_afterCallDestructors () at checks.cpp:450
#5  0x00000000004b5aab in check_parallel () at checks.cpp:204
#6  0x00000000004bba4d in runPass (tracker=..., passIndex=20) at runpasses.cpp:2

and since today it changed to:

#3  0x000000000055afcc in insertReferenceTemps (call=0xfe4be0) at callDestructors.cpp:881
#4  0x000000000053c78a in insertAutoCopyDestroyForTaskArg (arg=0x2703350, fcall=0xf16010, fn=0x396c0c0, firstCall=true) at parallel.cpp:264
#5  0x000000000053cad0 in bundleArgs (fcall=0xf16010, baData=...) at parallel.cpp:305
#6  0x00000000005424d6 in passArgsToNestedFns (nestedFunctions=...) at parallel.cpp:1169
#7  0x0000000000541ed9 in parallel () at parallel.cpp:1116
#8  0x00000000004bdd3b in runPass (tracker=..., passIndex=20) at runpasses.cpp:2