Annotate the nights where NPB EP and IS performed better, used -E
I had misinterpretted the desire to use -E as a desire to use it nightly rather
than a desire to maintain default behavior but compare it to tweaked behavior.
Since there isn't a good way to compare them (our execenvs don't allow multiple
settings), decided to just leave the default for now.