Remove part of the change I made to these .perfkeys last Friday
It turns out that I didn't look as closely at my results when testing on Friday
as I should have.  Adding the additional .dat file declaration to the .perfkeys
overrode the previous declaration, so all the information from over the weekend
was written to the last one instead of half and half like I thought it would.
Oops.

There was some merging needed anyways, the old file had a column that has been
replaced by our performance run verification step.  So when I fixed the .dat
files so all the data would be present, I removed all previous references to
the verification field (they were all successes anyways) and merged the file
headers.  When the graphs get generated for tomorrow's run, everything should
be back to normal.

For those following along at home, the data from over the weekend seems to
indicate that we are not adjusting our problem size as it gets faster.  Still
a mystery as to what is causing the improvement over time, further analysis is
needed.