Make start_test more multi-locale friendly
For multi-locale testing on crays we always have to throw --nostdinredirect and
--launchertimeout <launcher>. This just makes start_test infer those values for
most situations to make our lives easier when running start_test.

Previously start_test would not be able to run any tests or would never
complete execution if you didn't throw these on a cray and it wasn't obvious
what was going wrong

Throw --nostdinredirect unless there is no launcher or the launcher is
amudprun.  we know that stdin redirection doesn't work for most launchers. only
do redirection when there's no launcher, or the launcher is amudprun, which we
know supports stdin redirection. Warn user about skipping tests unless they
threw -nostdinredirect or -stdinredirect or an env var is set to silence it

If the launcher name contains slurm set CHPL_LAUNCHER_TIMEOUT to slurm, and if
the launcher name contains pbs or qsub set CHPL_LAUNCHER_TIMEOUT to pbs. These
launchers handle timeouts themselves, and we don't need/want sub_test to be
trying to do it's own timeout mechanism.