Replace some stringMove() calls with string_copy_len().
Two of the three uses of stringMove() can use string_copy_len(), because the caller knows
that allocation needs to be performed.  This done, the interface to stringMove() can be
changed to never allocate; it just reuses the buffer provided.  That removes c_string_copy
from the interface and simplifies the implementation.

Also replaced the call to snprintf() in stringMove() with strncpy() instead.  There is no
reason to use formatted I/O there, and formatted I/O is somewhat more expensive.

Overall, this patch improves the efficiency of string assignment in NewString.chpl.