Reduce usage of PRIM_STRING_COPY.
The goal is to remove PRIM_STRING_COPY completely and replace it with extern calls.  We
can't do this quite yet: there is one call remaining in StringWriter that prevents this.

In other locations, used direct calls to string_copy() or string_copy_len() as
appropriate.

In NewString.chpl, added a null check to isEmptyString, since either a c_string of length
0 or a null c_string_copy can represent the empty string.