Don't use expect script for slurm-srun launcher
An expect script was used to launch interactive jobs for slurm. This was
because the slurm-gasnetrun_ibv launcher it was originally copied from also
used an expect script. I think the only reason that launcher used it is because
some others that it was pieced together from also used it. So far as I know the
only good reason to launch with an expect script is if the launcher will have
some real interaction with the user. e.g. ask them if they're sure they want to
continue even if they're not on the correct file system or something.

Using the expect script added a carriage return to all output lines which
screwed up testing diffs. This just removes the use of expect and replaces it
with a direct call to srun for interactive jobs. This makes no changes for
batch jobs.

With this change things like "start_test test/release/examples/hello.chpl" will
just work on slurm systems without needing to throw any extra arguments or
having to use the launchcmd