Strengthen check for bounded iterator and add zippered iter check
Need to be sure this optimization only fires for bounded iterators. Currently
it fired for bounded and unbounded. The check to do so should be strengthened
more. Also when looking at the iteratorExpr we looked for a tuple so that we
could replace ranges in a zippered iterator. However, this also fired for non
zippered iterators over tuples, which we didn't want to do.