Turn of warning when stdin redirection isn't supported
start_test auto detects when stdin redirection isn't supported and warned the
user that it was going to skip tests with stdin. The warning proved move
annoying than useful especially since it made things like 'make check' fail.
This removes the warning and just turns it into an 'Info' message. The user
won't see it reported at the end of test summary but that seems fine.