Skip test/io/ferguson/qio_test.test.c for Cygwin
This test depends on lseek to reset its position within a file,
and then reads a certain number of bytes to determine the file's
length.  However, on Cygwin, the use of lseek means that processes
reading text files can't depend on the number of bytes read to
determine their position within a file.  This test should be
rewritten for cygwin portability, and our qio runtime code checked
for similar assumptions.  Until then, it is not useful there.