Remove most references to graphInfo and divs in callbacks
We add the graphinfo and divs to the dygraph, but that can only be done after
the dygraph has been created. The graphInfo and divs are used in some of the
callbacks. I believe that in cases where the dygraph loads quickly the
divs/graphInfo are trying to be accessed before they are actually set which is
why we sometimes get missing graphs