Use json instead of csv for dygraphs
Instead of gengraphs generating a csv we now generate json. jquery is used to
load and parse the json which is passed to dygraphs as a native array.

This adds a function to gengraphs to convert the csv to json and add labels to
the graphinfo (since labels must be explicitly set when using dygraphs native
array format.)

Instead of taking a csv, dygraphs now takes the native array. Jquery is used to
parse the json, and then we date-ify the data since json doesn't have a way to
specify dates.

This is being done so that graph expansion can be simplified, sped up, and made
to work with multiple configurations (so compiler perf graphs can use
multi-conf.) It should also slightly speed up loading normal graphs since
dygraphs no longer has to parse a csv.

This is a stepping stone towards simplifying compiler performance, beefing up
multiple configuration support (to be used with compiler performance, --local
vs --no-local and the various comm layers for multi-locale performance.) It
will also allow graph expansion to be simplified and sped up so that we can add
an expand button to any graph