uniquify dygraph csv names
We had a STREAM.graph and stream.graph file. One of which overrode the other.
This makes the filename unique. This isn't a great means of making the name
unique.

I'll come back to it later (I ran into this with my other graph work) and add
a check or a better way to make it unique.  There should also probably be a
check to make sure the graph titles are unique.