Put dygraph labels in json file
Previously I was putting the lables in graphdata.js. Now it goes in the same
json file as the data does. This is so you can look at the labels and data in
the same file, and so that the website doesn't have to load as much for graphs
it doesn't display