Make perf graph keys cleaner
Strip '=', '-', ' ', and ':' from the right side of graphkeys. We get a lot of
these when perfkeys are used as the graphkeys. It makes the series look strange
in dygraphs. This just strips the characters in gengraphs.

I used to do this for multi-conf because text was added to the end and those
trailing characters made it look really odd. That no longer needs to be done.