Change name of _build_range to be more explicit
Splitting the _build_range functions into 2 functions: chpl_build_bounded_range
and chpl_build_unbounded_range. This is for the upcoming anonymous range
iteration optimization where we want a clean way to find fully bounded ranges
early in compilation.

Added a chpl_ prefix instead of keeping the single underscore since that's the
direction we want to move with internal functions.