Add annotations for regressions from 12/05, 11/09
Simon's library mode commit seems to have impacted LULESH and binary-trees.
This doesn't quite make sense, but it's definitely the commit.  I will
investigate further into why this changed and if the impact can be reversed.

We also forgot to note when Brad turned off the task table, improving
compilation and the size of the generated code