Need to add an error case for invalid types on partially bounded ranges
Since all the range builders used to have the same name, and the partially
bounded ranges and fully bounded ranges took the same number of parameters, the
generic error case applied to unbounded ranges as well. It no longer does since
the names are different so adding an error case.