Split up builders for unbounded ranges
Instead of having:
 - chpl_build_bounded_range
 - chpl_build_unbounded_range

we now have:
 - chpl_build_bounded_range
 - chpl_build_partially_bounded_range
 - chpl_build_unbounded_range

It seemed worthwhile to split the builders up further to be able to identify
partially bounded vs completely unbounded ranges. It doesn't feel worthwhile to
split up the partially bounded into low and high bounded (though it'd be easy
to do in the future if we change our minds.)

If you have a range like "1.." it feels more natural to call it a partially
bounded range instead of an unbounded range. I suppose it could be called a
partially unbounded range, but that feels strange. Unbounded implies ".." to
me, so it makes sense to split out the builders at the very least to avoid any
confusion.