Remove extraneous add to inits in function resolution.
This line was the result of a bad "cut" command.  It was part of a previous if
clause that got removed, and somehow that left it in a different if clause than
where it started.  I reviewed the commit that allowed this, I should've caught
it.  The add was harmless, as it would be dead code the second time it was
reached and so skipped, but still.