Rename STREAM .graphs in studies directory
The csv files we generate for dygraphs get their name from the lowecase
basename of the .graph files. We had some duplicate basenames for some stream
tests. This renames the .graph files in the studies directory so we no longer
have a conflict. Previously we would override the .dat files, so the wrong data
was being loaded for some graphs.

An upcoming patch will add a check to make sure we don't run into this again.