Update checkGraphFiles to look for duplicate .graph files
Updates checkGraphFiles to look for duplicate .graph files with non-unique case
insensitive names. If there are duplicate .graph files this causes issues for
performance graphs. The csv files we generate are named by using the lower case
basename of the graph files so they will override each other if we have
non-unique .graph files.

This updates checkGraphFiles to do this check, and largely rewrites it to use
better python idioms and to use python functions instead of calling out to
things like 'find'

Also updates the warning in start_test and runs the check when we're doing
compiler performance testing as well.