Make returnStarTuplesByRefArgs() put the return variable write-back in the right place.
I stumbled across this long-standing bug after disabling returnRecordsByRefArguments().
Apparently that routine covered at least some cases that must be handled by
returnStartTuplesByRefArgs() when it is disabled.

If there is an early return in the routine, it will jump to the exit label.  As written,
returnStarTuplesByRefArgs() was putting the write-back to the return variable before the
exit label.  This works fine for the straight-line path, but fails to write out the result
when the early return fires.  This results in the caller using uninitialized memory, which
is generally bad.

The fix is a one-liner.  The call to insertBeforeReturn() that inserts the write-back was
changed to insertBeforeReturnAfterLabel().  That puts the write-back in the right place.