Handle when there is no data
When there is no data for a particular graph, we need to create a placeholder.
Dygraphs doesn't accept an empty array for the data, so we create a single
entry for today's date with null for all the data. This wasn't an issue when we
had csv's because dygraphs handled that internally, but it does not when we
pass the array as a native array.