First, in Makefile.include, change the ordering of the make(1) variable definitions to be generally:   base and unique build dirs   build dirs   install dirs   source dirs and files, if any   lib dirs and files, if any
Second, these lines in Makefile were superfluous:

  GMP_ABS_DIR = $(shell pwd)
  GMP_INSTALL_DIR = $(GMP_ABS_DIR)/$(GMP_INSTALL_SUBDIR)
  GMP_BUILD_DIR = $(GMP_ABS_DIR)/$(GMP_BUILD_SUBDIR)
  GMP_DIR = $(GMP_ABS_DIR)

Because GMP_ABS_DIR isn't referenced anywhere else, they're equivalent
to:

  GMP_DIR = $(shell pwd)
  GMP_INSTALL_DIR = $(GMP_DIR)/$(GMP_INSTALL_SUBDIR)
  GMP_BUILD_DIR = $(GMP_DIR)/$(GMP_BUILD_SUBDIR)

But these latter two definitions of GMP_{INSTALL,BUILD}_DIR are
identical to the ones in Makefile.include, which is included here by
virtue of ../../make/Makefile.base being included at the top. So they
can go. And the first definition here is effectively the same as this
one that is already in Makefile.include:

  GMP_DIR = $(shell pwd)/../../third-party/qthread

So, it can go as well.

Assuming this creates no problems in nightly testing, I'll change the
other affected third-party packages in a follow-up commit.