Drop noperf tests, instead of skipping them.
When running performance tests, there are lots of tests that are skipped
because they do not have the performance configuration files. They should not
be included in the jUnit reports as they are not interesting data points. Real
performance tests that are skipped due to the environment/etc are still
included as skipped in the report.