Add a test on the use of --no-use-noinit with complexes
I also neglected to add a test for complexes here.  This is particularly
relevant, because the module definition of complexes requires noinit always be
honored, due to the implementation of type defaults in module code.  It is our
intention to alter the type default implementation, and in such a way that
noinit can be skipped in the modules when --no-use-noinit is through.  This
test ensures that complexes work before and after such a change takes place,
(it would be one of many tests that would fail if the implementation of
--no-use-noinit were broken, but its presence should indicate more quickly what
precisely has gone wrong).