Fix --walltime parsing for sub_test
When passing --walltime to the executable (for slurm and pbs) the string we
created was only valid for times under 100 minutes. After that it became
invalid. We never run into this in nightly testing, but I ran into it when
trying to see how long hpl release would take to finish for gasnet+mpi