Gather performance data for the release version of HPL with MAX_LOGICAL
To avoid cluttering the release directory with files that are only relevant to
those interested in the performance implications of different
CHPL_RT_NUM_THREADS_PER_LOCALE, a new directory named perfComp has been made in
test/studies/hpcc/HPL/.  This directory contains symlinks to the portions of the
test that do not need to change between the two versions (the source code, the
.good file, the perfkeys and HPCCProblemSize), with alterations made to the
.perfexecopts file such that the two versions will not overwrite each other's
data.  Since this test is only interesting for performance runs (and should
always match the behavior of the original, being a symlink), add a skipif which
will only run when the performance flags have been thrown (at the moment, this
hard-codes --fast and --static into the compopts).

Additionally, add a suite to the performance graphs to display all tests which
are run with a different CHPL_RT_NUM_THREADS_PER_LOCALE value, in comparison to
their default runs.  This suite will soon contain more than just HPL.